extract FindspotRepository
