+++ src/fragmentarium/application/FindspotService.ts
@@ -1,20 +1,6 @@
 import Bluebird from 'bluebird'
-import { Findspot, fromFindspotDto } from 'fragmentarium/domain/archaeology'
-import { JsonApiClient } from 'index'
-
-export interface FindspotRepository {
-  fetchFindspots(): Bluebird<Findspot[]>
-}
-
-export class ApiFindspotRepository implements FindspotRepository {
-  constructor(private readonly apiClient: JsonApiClient) {}
-
-  fetchFindspots(): Bluebird<Findspot[]> {
-    return this.apiClient
-      .fetchJson('/findspots', false)
-      .then((findspots) => findspots.map(fromFindspotDto))
-  }
-}
+import { Findspot } from 'fragmentarium/domain/archaeology'
+import { FindspotRepository } from 'fragmentarium/infrastructure/FindspotRepository'
 
 export class FindspotService {
   constructor(private readonly findspotRepository: FindspotRepository) {}
